third_party/blink/web_tests/wpt_internal/html/canvas/placeElement/placeElement-100-percent-width-height.html

<!DOCTYPE html>

<head>
  <title>CanvasRenderingContext2D.PlaceElement() with 100% width and height fills parent container.</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <link rel="author" title="Aaron Krajeski" href="mailto:aaronhk@chromium.org" />
  <link rel="help" href="https://github.com/WICG/canvas-place-element" />
  <meta name="assert" content="Canvas fallback content fills the canvas with width/height = 100%" />
  <style>
    #fallback {
      background-color: green;
      width: 100%;
      height: 100%;
    }
  </style>
</head>

<body>
  <canvas>
    <div id="fallback">Placed Element</div>
  </canvas>
</body>

<script>
  const canvas = document.querySelector("canvas");
  const ctx = canvas.getContext("2d");
  ctx.placeElement(document.getElementById("fallback"), 0, 0);

  const rect = document.getElementById("fallback").getBoundingClientRect();
  test(function (t) {
    assert_equals(canvas.width, rect.width);
    assert_equals(canvas.height, rect.height);
  }, "Verify that the dimensions of the fallback div are correct.");
</script>
